pyproject.toml
[project]
name = "convert2pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "comtypes==1.4.1",
    "docx2pdf==0.1.8",
    "flask==3.0.0",
    "openpyxl==3.1.2",
    "pandas==2.1.4",
    "pdf2docx==0.5.6",
    "pdf2image==1.16.3",
    "pdfkit==1.0.0",
    "pillow==10.1.0",
    "pypdf2==3.0.1",
    "python-pptx==0.6.23",
    "reportlab==4.0.7",
    "tabula-py==2.9.0",
    "werkzeug==3.0.1",
]
